Source code and the database
The data base of all the users, the cost of the books for each branch
and for each semester is under the directory called DB.
This project needs some more work. If the password is entered wrongly,
the program has to be restarted. This aspects needs some working on. I
was too lazy to encrypt the passwords (more like i did not know then).
That needs work.
If possible one put it on a server and run it.